refactor: move catalogManager to its own folder and delete its duplicate
