+++ src/spa/logic/catalog/catalogDataManager/types.ts
@@ -4,6 +4,10 @@ import { ProductProjection } from '@commercetools/platform-sdk';
 export type CatalogData = {
   allCategories: Category[];
   allProducts: ProductProjection[];
+  categories: Category[];
+  categoriesThree: Record<string, Category[]>;
+  categoriesThreeText: Record<string, string[]>;
+  attributesArr: string[];
 };
 
 export interface CustomProductData {
@@ -15,3 +19,12 @@ export interface CustomProductData {
   discountPrice: string | null;
   imgURLs: string[];
 }
+
+export type IAllFiltersValue = {
+  brand: string;
+  color: string;
+  rangePrice: number[];
+  sale: boolean;
+  sortAlphabet: string;
+  sortPrice: string;
+};
