+++ src/spa/view/pages/catalogPage/catalogPageView.ts
@@ -3,12 +3,12 @@ import PageView from '@src/spa/view/pages/pageView';
 import { PAGE_NAME_ATTRIBUTE, PageNames } from '@src/spa/view/pages/types';
 import ElementCreator from '@src/spa/utils/elementCreator/elementCreator';
 import * as constants from '@src/spa/view/pages/catalogPage/constants';
-import { CatalogData } from '@src/spa/logic/catalog/types';
+import { CatalogData } from '@src/spa/logic/catalog/catalogDataManager/types';
 import { ProductProjection } from '@commercetools/platform-sdk';
-import CatalogDataManager from '@src/spa/logic/catalog/catalogDataManager';
+import CatalogDataManager from '@src/spa/logic/catalog/catalogDataManager/catalogDataManager';
 import SelectAttributeView from './select/selectAttribute';
 import InputView from '../../input/inputView';
-import { IAllFiltersValue } from '@src/spa/logic/catalog/types';
+import { IAllFiltersValue } from '@src/spa/logic/catalog/catalogDataManager/types';
 import CardProductView from './cardProduct/cardProduct';
 import { IRouter } from '@src/spa/logic/router/types';
 import { ElementCreatorParams, IElementCreator } from '@src/spa/utils/elementCreator/types';
